requirements.txt
# AI Coach Requirements - Production Ready
openai>=1.3.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Database support
psycopg2-binary>=2.9.0  # PostgreSQL adapter för Python
sqlalchemy>=2.0.0      # ORM för databas-operationer

# Security and Authentication
bcrypt>=4.0.0          # Säker lösenordshashing
cryptography>=41.0.0   # Kryptografi för tokens
email-validator>=2.0.0 # Email validering
passlib>=1.7.4         # Lösenordshantering

# Additional production dependencies
watchdog>=3.0.0  # För filövervakning
packaging>=21.0  # För kompatibilitet
uuid>=1.30       # För säkra session tokens